a2p-spam-filter/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
pyyaml
tldextract
python-multipart
pydantic>=2
requests
